Move UniqueTaskList into tasklist package
